% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{countBiologicalIDs}
\alias{countBiologicalIDs}
\title{Count biological IDs for InChIKey}
\usage{
countBiologicalIDs(inchikey)
}
\arguments{
\item{inchikey}{Query InChIKey. String}
}
\description{
Count biological IDs for InChIKey
}
\keyword{ID}
\keyword{biological,}
