% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{expandChemicalFormula}
\alias{expandChemicalFormula}
\title{Expand compact chemical formula}
\usage{
expandChemicalFormula(formula)
}
\arguments{
\item{formula}{Formula to be expanded. String.}
}
\value{
string. Expanded chemical formula
}
\description{
Expands compact chemical formula. Eg H20 to HHO
}
\examples{

}
\keyword{chemical,}
\keyword{expand,}
\keyword{formula}
