% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{InChIKey2MolDef}
\alias{InChIKey2MolDef}
\title{Convert InChIKey to molecule definition (MDL/SDF)}
\usage{
InChIKey2MolDef(inchikey)
}
\arguments{
\item{inchikey}{Query InChIKey. String}
}
\description{
Convert InChIKey to molecule definition (MDL/SDF)
}
\keyword{defininition}
\keyword{molecule,}
