% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{scoredInCHIKeys}
\alias{scoredInCHIKeys}
\title{Get a list of scored InChIKeys}
\usage{
scoredInCHIKeys(from, query, algorithm)
}
\arguments{
\item{from}{Source ID type. String.}

\item{algorithm}{Algorithm to use. Either 'biological' or 'popularity'}

\item{to}{Target ID type. String}
}
\description{
Get a list of scored InChIKeys
}
\keyword{defininition}
\keyword{molecule,}
