% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{sourceIDs}
\alias{sourceIDs}
\title{Obtain Valid Source IDs}
\usage{
sourceIDs()
}
\value{
list of source IDs
}
\description{
Produce list of valid source IDs that can be translated from.
}
\examples{
sourceIDs
}
\keyword{source}
\keyword{translate,}
