% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{targetIDs}
\alias{targetIDs}
\title{Obtain Valid Source IDs}
\usage{
targetIDs()
}
\value{
list of target IDs
}
\description{
Produce list of valid source IDs that can be translated to.
}
\examples{
fromVals
targetIDs
}
\keyword{target}
\keyword{translate,}
