% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugtranslator.R
\name{translateDrugID}
\alias{translateDrugID}
\title{Translate between identifiers}
\usage{
translateDrugID(from, to, query)
}
\arguments{
\item{from}{Source format. string. Valid list can be obtained by calling sourceIDs}

\item{to}{Target format. string. Valid list can be obtained by calling targetIDs}

\item{query}{Identifier to be translated. string.}
}
\value{
Target identifier
}
\description{
Translates drug/chemical IDs between a variety of standard
}
\examples{

}
\keyword{translate}
